RESTRUCTURED: source code and tests
This commit goes with restructured source code and unit tests.

SPLIT: Request unit tests.

RESTRUCTURED: The API Client source code.
